Add work for May 2019
Wil did a session on creating a REST server with Express. It got the
attendees really involved and sparked a lot of questions, which was good
as it allowed the mentors to really help folks out, which was definitely
rewarding, being able to help people out. But it also means that what's
in this directory is incomplete as I wasn't able to follow along as
much.