Add June 2019 session (Azure functions, web UI)
Nodeschool Seattle for June 1, 2019 was Brian Holt leading a session in
which we created an Azure function to access a third party API, and then
connected a create-react-app based UI to access it. Calling it a UI is
a little generous, since it's literally what is created when you run
create-react-app, with one div changed.

Main concepts for this session were:
* Azure functions (pretty cool, easy to get started with)
* Azure core tools (brew installed)
* VS Code Azure plugins

Note that we did not actually deploy the function to Azure, but ran it
locally through the help of the Azure core tools. It's pretty cool that
they can be run locally so easily. It'd be nice if the langauge support
were a little richer, but Go is in preview.